+++ tests/files/src/0607_internal_property_visibility.php
@@ -1,0 +1,6 @@
+<?php
+
+class MyException extends Exception {
+    protected $message = 'default';
+    private $code = 7;  // not visible enough
+}